+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add page for a big loyalty card
 - Add loyalty cards deletion
 - Add password changing
+- Add item picked picture
 
 ### Changed